requirements.txt
streamlit
pandas
numpy
seaborn
plotly
python-docx
pytesseract
PyPDF2
pydantic
python-dotenv
google-generativeai